requirements.txt
streamlit
pillow
google-generativeai
python-dotenv
langchain_community
langchain